+++ Wire-iOS/Sources/UserInterface/Conversation/InputBar/MentionsHandler+TextView.swift
@@ -18,15 +18,6 @@
 
 import Foundation
 
-fileprivate extension NSAttributedString {
-    func hasSpaceAt(position: Int) -> Bool {
-        guard wholeRange.contains(position) else { return false }
-        let scalars = attributedSubstring(from: NSRange(location: position, length: 1)).string.unicodeScalars
-        let justSpaces = scalars.filter(NSCharacterSet.whitespacesAndNewlines.contains)
-        return !justSpaces.isEmpty
-    }
-}
-
 extension MentionsHandler {
 
     static func cursorPosition(in textView: UITextView, range: UITextRange? = nil) -> Int? {
